Updated FFmpeg
git-svn-id: https://ffdshow-tryout.svn.sourceforge.net/svnroot/ffdshow-tryout/trunk@3317 3b938f2f-1a1a-0410-8054-a526ea5ff92c

+++ src/ffmpeg/libavutil/libm.h
@@ -92,7 +92,7 @@ static av_always_inline av_const float truncf(float x)
 #ifndef cbrtf
 static float cbrtf(float x)
 {
-    return pow(x, 1.0/3);
+    return pow((float)x, (float)1.0/3);
 }
 #endif
